ofproto-dpif-upcall: Attribute statistics to OpenFlow flows more precisely.
OpenFlow packet and byte counters have always been something of an
approximation in Open vSwitch.  First, they are updated only periodically.
Second, they can be misattributed because statistics collection does a
retranslation and gives the statistics to whichever OpenFlow flow or flows
they happen to hit, which could be different from the original set of
flows because the OpenFlow flow table may have changed.

This commit is intended to somewhat improve on the second point, by always
attributing statistics to the set of flows that were cached at the datapath
flow's initial translation by the revalidator.  This reduces the race in
which the flow table could have changed to the interval between the initial
installation and the first translation by the revalidator.  We plan to
start having the thread that does the initial installation also create the
ukey, which would eliminate that interval entirely.  At that point, packets
could get lost from OpenFlow statistics because of the periodic collection
of datapath statistics, but they should no longer be misattributed.

Signed-off-by: Ben Pfaff <blp@nicira.com>