ovsdb: add json cache
Although multiple jsonrpc monitor can share the same ovsdb monitor, each
change still needs to transalted into json object from scratch. This
can be wastful is mutiple jsonrpc monitors are interested in the
same changes.

Json chche impoves this by keeping an copy of json object generated
for transaction X to current transaction. When jsonrpc is interested
in a change, the cache is lookup first, if an json object is found,
a copy of it is handed back, skipping the regeneration process.

Any commit to the monitor will empty the cache. since none of them
will be useful anymore.

Signed-off-by: Andy Zhou <azhou@nicira.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>