[RFC] flow: Use new miniflow_builder for miniflow_extract().
This replaces the very specialized and difficult-to-maintain miniflow
assembly macros by a simpler data structure named "miniflow_builder" that
is more general and easier to use.

Unfortunately, in its current form it causes a significant performance
regression in miniflow_extract(), as measured by modifying test-flows.c
to run flow_extract 1,000,000 times in a loop (instead of just once).
Thus, I'm shelving this for now.

CC: Jarno Rajahalme <jrajahalme@nicira.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>